# src/test/resources/api/openapi.yaml
openapi: 3.1.0
info:
  title: api-datawarehousereader
  contact: {}
  license:
    name: MIT License
    url: https://opensource.org/licenses/MIT
  version: "5.0"
servers:
- url: http://localhost:50132
  description: Generated server url
...
paths:
  /{municipalityId}/measurements/{category}/{aggregateOn}:
    get:
      tags:
      - Measurement
      summary: Get measurement information
      description: Resource returns measurement data matching provided search parameters
      operationId: getMeasurements
      parameters:
      - name: municipalityId
        in: path
        description: Municipality id
        required: true
        schema:
          type: string
        example: 2281
      - name: category
        in: path
        required: true
        schema:
          $ref: "#/components/schemas/Category"
      - name: aggregateOn
        in: path
        required: true
        schema:
          $ref: "#/components/schemas/Aggregation"
      - name: partyId
        in: query
        description: PartyId (e.g. a personId or an organizationId)
        required: false
        schema:
          type: string
          description: PartyId (e.g. a personId or an organizationId)
          examples:
          - 81471222-5798-11e9-ae24-57fa13b361e1
      - name: facilityId
        in: query
        description: Facility id
        required: false
        schema:
          type: string
          description: Facility id
          examples:
          - "735999109151401011"
      - name: fromDateTime
        in: query
        description: "Date and time for oldest measurement point to return. Format\
          \ is yyyy-MM-dd'T'HH:mm:ss.SSSXXX, for example '2000-10-31T01:30:00.000-05:00'"
        required: false
        schema:
          type: string
          format: date-time
          description: "Date and time for oldest measurement point to return. Format\
            \ is yyyy-MM-dd'T'HH:mm:ss.SSSXXX, for example '2000-10-31T01:30:00.000-05:00'"
      - name: toDateTime
        in: query
        description: "Date and time for youngest measurement point to return. Format\
          \ is yyyy-MM-dd'T'HH:mm:ss.SSSXXX, for example '2000-10-31T01:30:00.000-05:00'"
        required: false
        schema:
          type: string
          format: date-time
          description: "Date and time for youngest measurement point to return. Format\
            \ is yyyy-MM-dd'T'HH:mm:ss.SSSXXX, for example '2000-10-31T01:30:00.000-05:00'"
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Measurement"
            application/problem+json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Measurement"
        "400":
          description: Bad request
          content:
            application/problem+json:
              schema:
                oneOf:
                - $ref: "#/components/schemas/Problem"
                - $ref: "#/components/schemas/ConstraintViolationProblem"
        "500":
          description: Internal Server error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
...
components:
  schemas:
    Problem:
      type: object
      properties:
        title:
          type: string
        detail:
          type: string
        status:
          $ref: "#/components/schemas/StatusType"
        instance:
          type: string
          format: uri
        type:
          type: string
          format: uri
        parameters:
          type: object
          additionalProperties: {}
    StatusType:
      type: object
      properties:
        statusCode:
          type: integer
          format: int32
        reasonPhrase:
          type: string
    ConstraintViolationProblem:
      type: object
      properties:
        cause:
          $ref: "#/components/schemas/ThrowableProblem"
        stackTrace:
          type: array
          items:
            type: object
            properties:
              classLoaderName:
                type: string
              moduleName:
                type: string
              moduleVersion:
                type: string
              methodName:
                type: string
              fileName:
                type: string
              lineNumber:
                type: integer
                format: int32
              className:
                type: string
              nativeMethod:
                type: boolean
        type:
          type: string
          format: uri
        status:
          $ref: "#/components/schemas/StatusType"
        violations:
          type: array
          items:
            $ref: "#/components/schemas/Violation"
        title:
          type: string
        message:
          type: string
        detail:
          type: string
        instance:
          type: string
          format: uri
        parameters:
          type: object
          additionalProperties: {}
        suppressed:
          type: array
          items:
            type: object
            properties:
              stackTrace:
                type: array
                items:
                  type: object
                  properties:
                    classLoaderName:
                      type: string
                    moduleName:
                      type: string
                    moduleVersion:
                      type: string
                    methodName:
                      type: string
                    fileName:
                      type: string
                    lineNumber:
                      type: integer
                      format: int32
                    className:
                      type: string
                    nativeMethod:
                      type: boolean
              message:
                type: string
              localizedMessage:
                type: string
        localizedMessage:
          type: string
    ThrowableProblem:
      type: object
      properties:
        cause: {}
        stackTrace:
          type: array
          items:
            type: object
            properties:
              classLoaderName:
                type: string
              moduleName:
                type: string
              moduleVersion:
                type: string
              methodName:
                type: string
              fileName:
                type: string
              lineNumber:
                type: integer
                format: int32
              className:
                type: string
              nativeMethod:
                type: boolean
        message:
          type: string
        title:
          type: string
        detail:
          type: string
        status:
          $ref: "#/components/schemas/StatusType"
        instance:
          type: string
          format: uri
        type:
          type: string
          format: uri
        parameters:
          type: object
          additionalProperties: {}
        suppressed:
          type: array
          items:
            type: object
            properties:
              stackTrace:
                type: array
                items:
                  type: object
                  properties:
                    classLoaderName:
                      type: string
                    moduleName:
                      type: string
                    moduleVersion:
                      type: string
                    methodName:
                      type: string
                    fileName:
                      type: string
                    lineNumber:
                      type: integer
                      format: int32
                    className:
                      type: string
                    nativeMethod:
                      type: boolean
              message:
                type: string
              localizedMessage:
                type: string
        localizedMessage:
          type: string
    Violation:
      type: object
      properties:
        field:
          type: string
        message:
          type: string
    Category:
      type: string
      description: Category
      enum:
      - COMMUNICATION
      - DISTRICT_COOLING
      - DISTRICT_HEATING
      - ELECTRICITY
      - ELECTRICITY_TRADE
      - WASTE_MANAGEMENT
      - WATER
    Aggregation:
      type: string
      description: Data point aggregation granularity
      enum:
      - QUARTER
      - HOUR
      - DAY
      - MONTH
    Measurement:
      type: object
      description: Measurement model
      properties:
        uuid:
          type: string
          description: Unique identifier for the measurement
          examples:
          - 550e8400-e29b-41d4-a716-446655440000
          readOnly: true
        customerOrgId:
          type: string
          description: Customer organization identifier
          examples:
          - "5534567890"
          readOnly: true
        facilityId:
          type: string
          description: Facility identifier
          examples:
          - "735999109151401011"
          readOnly: true
        feedType:
          type: string
          description: Type of feed
          examples:
          - Energy
          readOnly: true
        unit:
          type: string
          description: Measurement unit
          examples:
          - kWh
          readOnly: true
        usage:
          type: number
          description: Usage value
          examples:
          - "1500"
          readOnly: true
        interpolation:
          type: integer
          format: int32
          description: Interpolation value indicating data quality
          examples:
          - "0"
          readOnly: true
        dateAndTime:
          type: string
          format: date-time
          description: Date and time of the measurement
          examples:
          - 2024-01-15T10:30:00Z
          readOnly: true
...
  securitySchemes: {}